e2e/requirements.txt
requests>=2.22.0
PyYaml
python-gitlab
